refactor(header): extract useHideHeaderOnScroll hook
